requirements.txt
canonicalwebteam.flask-base==1.1.0
django-openid-auth==0.17
Flask-OpenID==1.3.0
alembic==1.13.1
psycopg2-binary==2.9.6
SQLAlchemy==2.0.25
Flask-Admin==1.6.1
python-slugify==8.0.4
Flask-WTF==1.2.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
